Rename interfaces from *Controller to *Api
