+++ modules/cluster-management/src/integrationTest/java/org/apache/ignite/internal/cluster/management/rest/ItClusterManagementControllerTest.java
@@ -44,7 +44,7 @@ import java.util.List;
 import org.apache.ignite.internal.cluster.management.ClusterManagementGroupManager;
 import org.apache.ignite.internal.cluster.management.MockNode;
 import org.apache.ignite.internal.rest.api.Problem;
-import org.apache.ignite.internal.rest.api.cluster.ClusterManagementController;
+import org.apache.ignite.internal.rest.api.cluster.ClusterManagementApi;
 import org.apache.ignite.internal.rest.api.cluster.ClusterStateDto;
 import org.apache.ignite.internal.testframework.WorkDirectory;
 import org.apache.ignite.internal.testframework.WorkDirectoryExtension;
@@ -113,7 +113,7 @@ public class ItClusterManagementControllerTest {
 
     @Test
     void testControllerLoaded() {
-        assertNotNull(server.getApplicationContext().getBean(ClusterManagementController.class));
+        assertNotNull(server.getApplicationContext().getBean(ClusterManagementApi.class));
     }
 
     @Test
