+++ modules/rest-api/src/main/java/org/apache/ignite/internal/rest/api/configuration/ClusterConfigurationApi.java
@@ -35,28 +35,27 @@ import org.apache.ignite.internal.rest.api.Problem;
 import org.apache.ignite.internal.rest.constants.MediaType;
 
 /**
- * Node configuration controller.
+ * Cluster configuration controller.
  */
-@Controller("/management/v1/configuration/node")
-@Tag(name = "nodeConfiguration")
-public interface NodeConfigurationController {
+@Controller("/management/v1/configuration/cluster/")
+@Tag(name = "clusterConfiguration")
+public interface ClusterConfigurationApi {
 
     /**
-     * Returns node configuration in HOCON format. This is represented as a plain text.
+     * Returns cluster configuration in HOCON format. This is represented as a plain text.
      *
-     * @return the whole node configuration in HOCON format.
+     * @return the whole cluster configuration in HOCON format.
      */
-    @Operation(operationId = "getNodeConfiguration")
+    @Operation(operationId = "getClusterConfiguration")
     @ApiResponses({
-            @ApiResponse(responseCode = "200",
-                    content = @Content(mediaType = MediaType.TEXT_PLAIN,
-                            schema = @Schema(type = "string")),
-                    description = "Whole node configuration"),
+            @ApiResponse(
+                    responseCode = "200",
+                    content = @Content(mediaType = MediaType.TEXT_PLAIN, schema = @Schema(type = "string")),
+                    description = "Get cluster configuration"),
             @ApiResponse(responseCode = "500", description = "Internal error",
                     content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class))),
             @ApiResponse(responseCode = "400", description = "Incorrect configuration",
                     content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class)))
-
     })
     @Produces({
             MediaType.TEXT_PLAIN, // todo: IGNITE-17082
@@ -66,22 +65,38 @@ public interface NodeConfigurationController {
     String getConfiguration();
 
     /**
+     * Updates cluster configuration in HOCON format. This is represented as a plain text.
+     *
+     * @param updatedConfiguration the cluster configuration to update.
+     */
+    @Operation(operationId = "updateClusterConfiguration")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "Configuration updated"),
+            @ApiResponse(responseCode = "500", description = "Internal error",
+                    content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class))),
+            @ApiResponse(responseCode = "400", description = "Incorrect configuration",
+                    content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class)))
+    })
+    @Consumes(MediaType.TEXT_PLAIN) // todo: IGNITE-17082
+    @Produces(MediaType.PROBLEM_JSON)
+    @Patch
+    CompletableFuture<Void> updateConfiguration(@Body String updatedConfiguration);
+
+    /**
      * Returns configuration in HOCON format represented by path. This is represented as a plain text.
      *
-     * @param path to represent a node configuration.
+     * @param path to represent a cluster configuration.
      * @return system configuration in HOCON format represented by given path.
      */
-    @Operation(operationId = "getNodeConfigurationByPath")
+    @Operation(operationId = "getClusterConfigurationByPath")
     @ApiResponses({
             @ApiResponse(responseCode = "200",
-                    content = @Content(mediaType = MediaType.TEXT_PLAIN,
-                            schema = @Schema(type = "string")),
+                    content = @Content(mediaType = MediaType.TEXT_PLAIN, schema = @Schema(type = "string")),
                     description = "Configuration represented by path"),
             @ApiResponse(responseCode = "500", description = "Internal error",
                     content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class))),
             @ApiResponse(responseCode = "400", description = "Incorrect configuration",
                     content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class)))
-
     })
     @Produces({
             MediaType.TEXT_PLAIN, // todo: IGNITE-17082
@@ -89,23 +104,4 @@ public interface NodeConfigurationController {
     })
     @Get("/{path}")
     String getConfigurationByPath(@PathVariable("path") String path);
-
-    /**
-     * Updates node configuration in HOCON format. This is represented as a plain text.
-     *
-     * @param updatedConfiguration the node configuration to update. This is represented as a plain text.
-     */
-    @Operation(operationId = "updateNodeConfiguration")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "Configuration updated"),
-            @ApiResponse(responseCode = "500", description = "Internal error",
-                    content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class))),
-            @ApiResponse(responseCode = "400", description = "Incorrect configuration",
-                    content = @Content(mediaType = MediaType.PROBLEM_JSON, schema = @Schema(implementation = Problem.class)))
-
-    })
-    @Consumes(MediaType.TEXT_PLAIN) // todo: IGNITE-17082
-    @Produces(MediaType.PROBLEM_JSON)
-    @Patch
-    CompletableFuture<Void> updateConfiguration(@Body String updatedConfiguration);
 }
