+++ modules/rest/src/main/java/org/apache/ignite/internal/rest/RestComponent.java
@@ -34,10 +34,10 @@ import org.apache.ignite.configuration.schemas.rest.RestView;
 import org.apache.ignite.internal.logger.IgniteLogger;
 import org.apache.ignite.internal.logger.Loggers;
 import org.apache.ignite.internal.manager.IgniteComponent;
-import org.apache.ignite.internal.rest.api.cluster.ClusterManagementController;
-import org.apache.ignite.internal.rest.api.configuration.ClusterConfigurationController;
-import org.apache.ignite.internal.rest.api.configuration.NodeConfigurationController;
-import org.apache.ignite.internal.rest.api.node.NodeManagementController;
+import org.apache.ignite.internal.rest.api.cluster.ClusterManagementApi;
+import org.apache.ignite.internal.rest.api.configuration.ClusterConfigurationApi;
+import org.apache.ignite.internal.rest.api.configuration.NodeConfigurationApi;
+import org.apache.ignite.internal.rest.api.node.NodeManagementApi;
 import org.apache.ignite.lang.IgniteInternalException;
 import org.jetbrains.annotations.Nullable;
 
@@ -53,10 +53,10 @@ import org.jetbrains.annotations.Nullable;
         license = @License(name = "Apache 2.0", url = "https://ignite.apache.org"),
         contact = @Contact(email = "user@ignite.apache.org")))
 @OpenAPIInclude(classes = {
-        ClusterConfigurationController.class,
-        NodeConfigurationController.class,
-        ClusterManagementController.class,
-        NodeManagementController.class
+        ClusterConfigurationApi.class,
+        NodeConfigurationApi.class,
+        ClusterManagementApi.class,
+        NodeManagementApi.class
 })
 public class RestComponent implements IgniteComponent {
     /** Default port. */
@@ -142,7 +142,6 @@ public class RestComponent implements IgniteComponent {
     }
 
     private void setFactories(Micronaut micronaut) {
-        micronaut.classes(NodeManagementController.class);
         for (var factory : restFactories) {
             micronaut.singletons(factory);
         }
